+++ internal/app/go-filecoin/storage_miner_connector/connector.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/connector_common"
+	"github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/connectors"
 
 	"github.com/filecoin-project/go-address"
 	storagenode "github.com/filecoin-project/go-storage-miner/apis/node"
@@ -320,7 +320,7 @@ func (m *StorageMinerNodeConnector) GetSealTicket(ctx context.Context, tok stora
 }
 
 func (m *StorageMinerNodeConnector) GetChainHead(ctx context.Context) (storagenode.TipSetToken, abi.ChainEpoch, error) {
-	return connector_common.GetChainHead(m.chainState)
+	return connectors.GetChainHead(m.chainState)
 }
 
 // GetSealSeed is used to acquire the interactive seal seed for the provided pre-commit
